add GenerateAndSendOrderListQueue
